Fix formatting and type errors across codebase
Apply ruff formatting to 15 files and fix 4 type errors caught by ty:
- handler.py: use Callable type instead of builtin callable
- protocol.py: use isinstance for proper type narrowing in route_request
- server.py: fix configure_logging kwarg name (level -> log_level)
- config.py: change rate_limit_capacity from float to int

+++ src/titlani/server/config.py
@@ -20,7 +20,7 @@ class ServerConfig:
 
     # Rate limiting
     rate_limit_enable: bool = False
-    rate_limit_capacity: float = 10.0
+    rate_limit_capacity: int = 10
     rate_limit_refill_rate: float = 1.0
     rate_limit_retry_after: int = 30
 
@@ -51,44 +51,26 @@ class ServerConfig:
         if "keyfile" in server:
             config.keyfile = Path(server["keyfile"])
         if "identity_certfile" in server:
-            config.identity_certfile = Path(
-                server["identity_certfile"]
-            )
+            config.identity_certfile = Path(server["identity_certfile"])
         if "identity_keyfile" in server:
             config.identity_keyfile = Path(server["identity_keyfile"])
 
         config.rate_limit_enable = rate_limit.get("enable", False)
-        config.rate_limit_capacity = rate_limit.get("capacity", 10.0)
-        config.rate_limit_refill_rate = rate_limit.get(
-            "refill_rate", 1.0
-        )
-        config.rate_limit_retry_after = rate_limit.get(
-            "retry_after", 30
-        )
+        config.rate_limit_capacity = int(rate_limit.get("capacity", 10))
+        config.rate_limit_refill_rate = rate_limit.get("refill_rate", 1.0)
+        config.rate_limit_retry_after = rate_limit.get("retry_after", 30)
 
-        config.access_control_enable = access_control.get(
-            "enable", False
-        )
-        config.access_control_allow_list = access_control.get(
-            "allow_list", []
-        )
-        config.access_control_deny_list = access_control.get(
-            "deny_list", []
-        )
-        config.access_control_default_allow = access_control.get(
-            "default_allow", True
-        )
+        config.access_control_enable = access_control.get("enable", False)
+        config.access_control_allow_list = access_control.get("allow_list", [])
+        config.access_control_deny_list = access_control.get("deny_list", [])
+        config.access_control_default_allow = access_control.get("default_allow", True)
 
         return config
 
     def validate(self) -> None:
         if self.certfile and not self.certfile.exists():
-            raise ValueError(
-                f"Certificate file not found: {self.certfile}"
-            )
+            raise ValueError(f"Certificate file not found: {self.certfile}")
         if self.keyfile and not self.keyfile.exists():
-            raise ValueError(
-                f"Key file not found: {self.keyfile}"
-            )
+            raise ValueError(f"Key file not found: {self.keyfile}")
         if self.port < 1 or self.port > 65535:
             raise ValueError(f"Invalid port: {self.port}")
